pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "levante-bench"
version = "0.1.0"
description = "LEVANTE VLM benchmark: compare VLMs to children's behavioral data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26,<3",
    "pandas>=2.2",
    "torch>=2.2",
    "transformers>=4.45",
    "Pillow>=10",
    "requests>=2.31",
    "omegaconf>=2.3",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "black", "isort"]

[project.scripts]
levante-bench = "levante_bench.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
